+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased] - 2019-05-02
+(Poikilos)
+### Changed
+* Attempt download in Python 3, but document lack of support.
+
 ## [Unreleased] - 2019-05-01
 (Poikilos)
 ### Changed